### 2019-09-03 #338 @yihanwu
baRcodeR Submission
---

Submitting Author: Yihan Wu (@yihanwu) and Robert Colautti (@ColauttiLab)
Repository: [yihanwu/baRcodeR](https://github.com/yihanwu/baRcodeR)
Version submitted:0.1.4
Version accepted: TBD

---

```
Package: baRcodeR
Title:  Label Creation for Tracking and Collecting Data from Biological Samples
Version: 0.1.4
    c(person(given = "Yihan",
             family = "Wu",
             role="aut",
             email = "yihan.wu@queensu.ca",
      person(given = "Robert",
             family = "Colautti",
             email = "robert.colautti@queensu.ca",
             comment = c(ORCID = "0000-0003-4213-0711")))
Maintainer: Yihan Wu <yihan.wu@queensu.ca>
Description: Tools to generate unique identifier codes and printable barcoded
    labels for the management of biological samples. The creation of unique ID
    codes and printable PDF files can be initiated by standard commands, user
    prompts, or through a GUI addin for R Studio. Biologically informative codes
    can be included for hierarchically structured sampling designs.
    R (>= 3.4.0),
    qrcode
License: file LICENSE
Language: en-CA
    DT (>= 0.3),
    grDevices,
    grid,
    miniUI (>= 0.1.1),
    shiny (>= 0.13),
    rstudioapi
    covr,
    shinytest

```

Reproducibility: baRcodeR facilitates the creation of unique text identifier (ID) codes and scannable barcodes (linear and 2D) that can be printed onto customized sticker labels with consumer-grade printers to facilitate management, tracking, and data collection involving biological samples. Unique identifier codes and scannable barcode labels are produced via reproducible code which reduce transcription errors that arise with hand-written labels and manual entry of label codes. In contrast to commercial software, no proprietary hardware is required, labels can be generated automatically with biologically informative coding, and labels can be created from the command line.

baRcodeR is designed to minimize bookkeeping errors and to assist researchers who need to track, curate, or measure biological samples. It is particularly well-suited for large collaborative projects with complex sampling design and numerous samples shared among researchers. For example, a large biological research project may have subjects organized as population/treatment-by-family/line-by-time – that is, each subject sampled multiple times and representing a specific family group or genetic line from one or more nested experimental populations or treatments. The custom layout options in baRcodeR allow researchers to rapidly generate meaningful identifier codes that capture the hierarchical sampling structure, and then print customizable labels containing the ID codes as human-readable text with corresponding digital linear or 2D barcodes. The command-line interface also facilitates rapid re-creation of a set of generated identifiers with different sized labels or minor changes to identifier codes – for example, to label multiple tissue collection containers and then later to label test tubes for purified extracts, each with the same basic code with a new code corresponding to sample type.

    - [zintr](https://github.com/carlganz/zintr) creates individual barcode image files but does not include functions for (i) automating the creation of biologically-relevant, unique ID codes or (ii) customizable layouts for printing multiple barcodes.
    - [qrcode](https://cran.r-project.org/web/packages/qrcode/index.html) generates qrcodes and is a dependency for baRcodeR to generate 2D barcodes. However, qrcodes also does not include functions for (i) automating the creation of biologically-relevant, unique ID codes or (ii) customizable layouts for printing multiple barcodes.

[Presubmission inquiry #336](https://github.com/ropensci/software-review/issues/336), approved by editor @melvidoni
