### 2019-09-04 #338 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @yihanwu  !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). You don't need to address these now, it's info for reviewers to use to get started.

```r
── GP baRcodeR ─────────────

It is good practice to

    R/hidden_createPDF.R:272:NA
    R/hidden_createPDF.R:307:NA
    R/hidden_createPDF.R:308:NA
    R/hidden_createPDF.R:309:NA
    R/hidden_createPDF.R:310:NA
    ... and 14 more lines

  ✖ write short and simple functions. These functions have high cyclomatic complexity:custom_create_PDF (58).
  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with other packages. Use "Imports" instead.
  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals 'TRUE' and 'FALSE' by default, but are not reserved words and hence can
    be overwritten by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the logicals.

    R/hidden_createPDF.R:NA:NA
 ───────────────
```

Seeking reviewers now 🕐

---

Reviewers:

- xx
- xx
