### 2019-09-10 #338 @raynamharris
- [ ] **Vignette(s)** demonstrating major functionality that runs successfully locally. _The vingnette did not load._

   and a reasonable range of inputs and conditions. All tests pass on the local machine. _Did not perform automated tests_

---

`baRcodeR` is an R package that reproducibly creates sample labels. It can be run from the RStudio console or from a GUI Addin. This is the first time I have ever used a GUI or Addin from within RStudio, and it worked very well _after_ restarting R twice. I would recommend this package to my colleagues who do molecular work. Below is a list of minor comments and concerns.

- In the README, there is a "See also:" link to a similar package called `zintr`, but it is unclear why a user might want to refer to it or why they should chose one over the other or both. A brief elaboration would be helpful. Also, it would be useful to state that Zint is a barcode generator and library.
- I really like the graphical overview and the cheatsheet. These are both very useful. I also like that your cheatsheet is archived in FigShare with a DOI.
- Does this default layout only work with Uline products or is there an Avery alternative?
- In order for the baRcodeR GUI to appear in my Addins window, I had to restart RStudio. Perhaps add a note that you made need to restart R after successful installation of the package.
- The first time I clicked on baRcodeR GUI my R session was aborted. It worked the second time, so maybe add a comment to restart and retry. Also, this is the message that accompanied the first, unsuccessful launch of the GUI and the aborted R session.

```
baRcodeR:::make_labels()
Registered S3 method overwritten by 'R.oo':
  method        from
  throw.default R.methodsS3
Loading required package: shiny
```

- The Simple ID Code maker works great!
- When I click "Create Lable.csv" it appeared like nothing happened, but the file was indeed saved in my working directory. A real bonus would be a pop-up window or a message printed to the console that said something like "file save in your working directory", but a simple fix would be to clarify this in the README.
- When following the instructions in the "Using the RStudio addin" section, the text explaining the process comes after the example image, much like a figure legend. However, it would be much easier to follow along if the text was above the image instead of below it. Also, there is very little white space between the images and the text. Some level three headers (eg. Simple ID Codes with `uniqID_maker` and Hierarcical ID codes with `uniqID_hier_maker` could help distinguish the sections.)
- When I click "add level", I keep getting the error message ``Label Preview Input list has only one level. Did you forget a level or are you sure you are not looking for uniqIDMaker()?`` I was, however, successfully able to make a hierarchical labels despite the error message. Is there a way to prevent this error message from appearing?
- Why is the default font size 2.2 in the GUI? This seems a little small. Also, in the manual, the default size is 5.
- I installed the package with `devtools::install_github("yihanwu/baRcodeR", build_vignettes = T, dependencies = T)` when when I run 1`vignette("Using-baRcodeR")`, I get the warning message `vignette ‘Using-baRcodeR’ not found`.
- I was able to past the code from the GUI into the command line and regenerate the same pdf. Success!
- I'm not sure I understand the error correction description. The L,M,Q,H makes me think it goes form low quality to high quality, but this is conflicts with the 7%-30% damage description. What does damage mean? What do LMQ and H stand for? Please clarify.
- I noticed that this repo has been forked from another repo ([nine2k/II-I-II-baRcodes-I-II-III](https://github.com/nine2k/II-I-II-baRcodes-I-II-III))? Do those authors know that this is being submitted as a package? Will they be offered co-authorship?

Here is the session info from my workflow.

```
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.5

BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib

[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

[1] stats     graphics  grDevices utils     datasets  methods
[7] base

[1] baRcodeR_0.1.4 qrcode_0.1.1   shiny_1.3.2

 [1] Rcpp_1.0.2        rstudioapi_0.10   knitr_1.24
 [4] magrittr_1.5      xtable_1.8-4      R6_2.4.0
 [7] rlang_0.4.0       stringr_1.4.0     tools_3.6.0
[10] grid_3.6.0        DT_0.8            xfun_0.9
[13] R.oo_1.22.0       miniUI_0.1.1.1    htmltools_0.3.6
[16] crosstalk_1.0.0   yaml_2.2.0        digest_0.6.20
[19] crayon_1.3.4      later_0.8.0       htmlwidgets_1.3
[22] R.utils_2.9.0     promises_1.0.1    rsconnect_0.8.13
[25] evaluate_0.14     mime_0.7          rmarkdown_1.15
[28] stringi_1.4.3     compiler_3.6.0    R.methodsS3_1.7.1
[31] jsonlite_1.6      httpuv_1.5.1
```
