### 2019-09-23 #338 @raynamharris
I was looking over the [rOpenSci peer review guidelines](https://devguide.ropensci.org/reviewerguide.html) today (because they are so good and I wanted to use this as a guide for a different peer review that I'm working on), and I came across this bullet point "If you have your own relevant data/problem, work through it with the package. You may find rough edges and use-cases the author didn’t think about."

In my current research, I always sample three specific biological tissue that that I code "hyp", "pit' and "gon". Is there a to make barcode with strings in addition to number? So, something like animal-01-hyp, animal-01-pit, animal-01-gon would be awesome.  If allowing any random string isn't possible, perhaps something,  "A", "B", and "C" could be added so that you could generate sample names like animal-01-A, animal-01-B, animal-01-C instead only only animal-01-tissue-01, animal-01-tissue-02, etc?

Does that make sense?
