### 2019-09-23 #338 @yihanwu
Hi @raynamharris,

It is possible to add a suffix string with the hierarchical labels.

For example, we can do,

```
animals <- c("animal", 1, 2)
tissues <- c("tissue", 1,3)
uniqID_hier_maker(hierarchy = list(animals, tissues), end = c("hyp", "pit", "gon"))
```

and the resulting data frame would look like:

||label|animal|tissue|
|---|---|---|---|
|1|animal01-tissue01hyp|animal01|tissue01|
|2|animal01-tissue02pit|animal01|tissue02|
|3|animal01-tissue03gon|animal01|tissue03|
|4|animal02-tissue01hyp|animal02|tissue01|
|5|animal02-tissue02pit|animal02|tissue02|
|6|animal02-tissue03gon|animal02|tissue03|

The original thought for the hierarchical labels was that tissue types could be coded as numbers but I can see how it would be overly complicated for cases like yours.

Just to confirm, you would want something that could generate:

```
animal01-hyp
animal01-pit
animal01-gon
animal02-hyp
animal02-pit
animal02-gon
animal03-hyp
 and so on
```

Here, each base label is repeated three times for the three ending strings.
