### 2019-09-28 #338 @llrs
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 7-8h

---

**Test installation**

First installation took long time (I don't have much of the dependencies of baRcodeR) but the dependencies look right to me.

The second install directly from github worked well and fast now that I already have the dependencies

**Check package integrity**

There is a note of non standard files found:

```
 Non-standard files/directories found at top level:
     ‘CODE_OF_CONDUCT.md’ ‘WuEtAl_2019_baRcodeR_MEE.pdf’ ‘docs’
```
The WuEtAl_2019_baRcodeR_MEE.pdf seems like an article about the program, which might be better on some preprint archive or another repository but it doesn't belong to the package. I would recommend to delete it from the repository.

CODE_OF_CONDUCT.md and the docs folder should be included in .Rbuildignore.
Also the docs folder is created to create the webpage of the package. Maybe you can automate this using `use_pkgdown_travis()` seeing that you already use Travis.

There are few tests, and 2 warnings related to introducing NAs by coercion (test_warnings.R, lines 12 and 13) and one about "Digits specified less than max number. Increasing number of digits." in test-label_generation.R line 13.

To test for warnings remember that there is the `expect_warning` function in testthat.
The other warning should be captured by `expect_warning` or select a good number to avoid the warning.

`uniqID_hier_maker` is only tested for warnings and not for expected output.

The shiny app is only tested locally. Maybe this was a recommendation from the CRAN?
I don't know in which platform it was tested but when I run it with
`compareImages = TRUE` it failed. If possible I would recommend to test it on CRAN even if it is with `compareImages = FALSE`.

It reports that the coverage is only 27% I would aim for 90% or more seeing that you want to test also the shiny app and the code.

It also warns about using `T` and `F` instead of `TRUE` and `FALSE`.

Also it recommends to think about the dependency to qrcodes that is on Depends instead of Imports. I think it is sensible to have it Depends as without it this packages seems useless.

However, `goodpractie::gp` couldn't continue with the analysis due to some error with linter. Probably this is due to some long string of code.

**Check of the metadata**

The README doesn't have any badge of status, from repostatus.org. See [this section](https://devguide.ropensci.org/building.html#code-style)

Most if not all of the results that the spelling check reports are from examples or names. Consider adding them the dictionary so that they are not reported again.

The readme recommends to provide the `sessionInfo`, you might be interested on the package [reprex](https://cran.r-project.org/package=reprex), which can help your users to create the minimal reproducible example.

The DESCRIPTION lists 6 Imports and 1 Depends but the NAMESPACE only imports two package.
The NAMESPACE should import fully the Depends packages and selectively the functions of packages on Imports.

It also uses Authors@R and Maintainer it is better to select just one (less things to keep up to date). Nice addition of the ORCID numbers.

I would recommen to use `usethis::use_tidy_description` to set the fields of the DESCRIPTION in a "standard" format.

**Check of the documentation**

The first thing I notice looking at the documentation is a link to another library.
I would move this reference below "Usage from the console".

I would also have a short section showing how you can use the library on the README.Rmd via the command line (without prompts).
This way the potential user get a grasp of what does the package.

Also probably it would be nice to show the final output of the package, currently it seems like the result is a plain data.frame.

The section in the README about "Using the RStudio addin" I think it worth to be in its own vignette.

There are some man pages (man/cheatsheet.Rd, man/make_labels.Rd) that call library or require on baRcodeR, it is not necessary.

It  is not recommended to call the system on a man page like in create_PDF.Rd. If you have to, use `system`, but `system2` would be better because the command is always quoted. So if you want to pass several commands you should call several times to `system2`.

There is no information about what do the `cheatsheet` and the `make_labels` return.
It is always good to tell your users what to expect from your functions.

There isn't any example on `custom_create_PDF`, I would recommend to add one even if they are not run automatically.

Also some examples do not run. From create_PDF:

```r
example_vector <- as.data.frame(c("ao1", "a02", "a03"))
create_PDF(Labels = example_vector, name = file.path(tempdir(), "example"))
# Warning message:
# In custom_create_PDF(user, Labels, name, type, ErrCorr, Fsz, ...) :
#   Cannot find a label column. Using first column as label input.
```
The details section of `create_PDF` has also a "#` " at the beginning.
The roxygen2 notation is duplicated on the source file, remove it and it will disappear

**Check the vignette**

The image of the workflow is very informative and useful for the user, but I would move it below the overview (as well as the cheatsheet).
As a user I would like to know what does the package first and then how can I do it.

The first code block of the vignette is to install some packages that are (or should be) automatically installed when the user installs baRcodeR.
The code block should be removed.

In the vignette you use `T` and `F` instead of `TRUE` and `FALSE`, `T` and `F` are not protected, so it might introduce errors, change them to the full name please.

Next you explain the workflow using prompts. Prompts are quoted, as they are printed in the terminal I would show them as code by using the three backticks formatting.

The next section is again using prompts. I was a bit confused about what is the goal of this section.
The columns of this hierarchical labels are the same of the text, which might be confusing.
Maybe the title could be one and the string used a different.

In the section "By argument" when I ran the code I got a warning:

```r
# Warning message:
# In uniqID_hier_maker(hierarchy = hier_list, digits = 1) :
#   Digits specified less than max level number. Increasing number of digits for level
```

It is mentioned again to store it in a csv. I tried to `create_PDF` and I didn't know where the file was created.
Apparently it is created on the working directory.
Maybe it should be better explained on the help page.
When I used the argument `user = TRUE` in `create_PDF` I got a prompt, I introduced the name of the pdf with quotes. Then the resulting file got also quotes. This could be checked and warned to the user as it could create problems.

**Check the functionality**

The functions are very well commented about what do they do.

Almost all the functions implement a sort of prompt.
I am not sure this is actually helpful as then the user need to know which question corresponds with which argument of the function.

The creation of labels seems a bit over-complicated to me.
We can obtain all the combinations of a group of variables with `expand.grid`

labels <- expand.grid(fixed = "a", variable = 1:5, join = "-",
                      text = letters[2:3], numbers = 7:9)
labels
apply(labels, 1, paste0, collapse = "")
```

This allows to omit the loop if you have a hierarchy.

The GUI addin is nice, but the "Simple ID Code Generation" and the "Hierarchical ID Code Generation" end up creating a labels.csv instead of creating the bar codes.
I expected that they would be directly created.
Also there isn't a way to select where the labels.csv is created.

I tried to create a Hierarchical ID Code Generation with the GUI and I end up with an error :
```
Error: Input list has only one level. Did you forget a level or are you sure you are not looking for uniqIDMaker()?
```
That's after creating a label and pressing "Add level".

**Inspection of the code**

Whenever possible avoid `sapply()`; use `vapply()`, as `sapply` will simplify the output which might cause errors. There are 8 cases of this, maybe some of them can be replaced, in some cases you just want to get the first element, maybe you can use `unlist` instead.

Also you use `1:...` like 5 times; use `seq_len()` or `seq_along()` or simply `seq()`.

In `code_128_make()` you used `T` and `F` replace them for `TRUE` and `FALSE`.

In hidden_createPDF.R you check a class with `class()`, it is better to use the `is()` function as it will show all the classes it inherits from.

There are very long functions (`make_labels_internals`, `custom_create_PDF` and `server` that are more than 100 lines) try splitting them in shorter functions. Usually functions of less of 50 lines are recommended.

There are also very long lines, consider shorter lines; 253 lines (15%) are > 80 characters long.

The `create_PDF` and the `custom_create_PDF` do the same.
I would delete `create_PDF`.
I understand that the idea was to avoid confusing the user with too many arguments.
But either do not show them or let them trust you with the defaults, which work reasonable well.

You make use of `grDevices::` and other packages but don't declare on the NAMESPACE that you import them.
You should do so with `#' @importFrom grDevices pdf` for instance.

**Check the testing**

The file `uniqID_maker_addin.R` is not tested.

I don't know how can you test when you interactively ask the user with prompts but that section should be tested too (hidden_createPDF.R).

**Conclusions**

Overall the package has its audience it is easy to use and provides a nice functionality for wet-dry researchers.
