### 2019-09-30 #338 @yihanwu
Thank you @llrs for the in-depth review!

I'll have a point to point review posted after I make the changes suggested by @llrs and @raynamharris
