### 2019-10-26 #338 @yihanwu
Thank you @raynamharris and @llrs for your detailed reviews and suggestions, which significantly improved the package. Below is a point-by-point response to issues raised, with added commit links where appropriate. Overall, we think we were able to address all of the significant concerns, including new test code to provide additional checks, a re-structured README with content on the add-in GUI moved to a separate vignette. We also added new functionality for making ID codes with ending strings as suggested by @raynamharris. (https://github.com/yihanwu/baRcodeR/commit/f550757a0ea58027d33a862cc980ea1be51da0ea)

## llrs review:

### Check package integrity

There is a note of non standard files found:

Non-standard files/directories found at top level:
     ‘CODE_OF_CONDUCT.md’ ‘WuEtAl_2019_baRcodeR_MEE.pdf’ ‘docs’
The WuEtAl_2019_baRcodeR_MEE.pdf seems like an article about the program, which might be better on some preprint archive or another repository but it doesn't belong to the package. I would recommend to delete it from the repository.

> 1. The manuscript has been deleted from the repository. (https://github.com/yihanwu/baRcodeR/commit/d0e53b9e1d182dd432e51b8cf4475f6e31f9309e)

CODE_OF_CONDUCT.md and the docs folder should be included in .Rbuildignore.
Also the docs folder is created to create the webpage of the package. Maybe you can automate this using use_pkgdown_travis() seeing that you already use Travis.

> 2. Updated .Rbuildignore for CODE_Of_CONDUCT.md and docs folder. (https://github.com/yihanwu/baRcodeR/commit/d0e53b9e1d182dd432e51b8cf4475f6e31f9309e)

There are few tests, and 2 warnings related to introducing NAs by coercion (test_warnings.R, lines 12 and 13) and one about "Digits specified less than max number. Increasing number of digits." in test-label_generation.R line 13.

To test for warnings remember that there is the expect_warning function in testthat.
The other warning should be captured by expect_warning or select a good number to avoid the warning.

> 3. Double warnings in test_warning and test-label_generation.R resolved (https://github.com/yihanwu/baRcodeR/commit/6bef6b993edc2bc55b13182e827296ad1d4831a3)

uniqID_hier_maker is only tested for warnings and not for expected output.

> 4. added a test for expected output for uniqID_hier_maker (https://github.com/yihanwu/baRcodeR/commit/6bef6b993edc2bc55b13182e827296ad1d4831a3)

The shiny app is only tested locally. Maybe this was a recommendation from the CRAN?
I don't know in which platform it was tested but when I run it with
compareImages = TRUE it failed. If possible I would recommend to test it on CRAN even if it is with compareImages = FALSE.

> 5. [shinytest](https://rstudio.github.io/shinytest/articles/ci.html) recommends using compareImages = FALSE because images rendered on different platforms are different. For example, the images generated on Travis (Linux) will be different from mine (Windows). They also advise not to [check on CRAN](https://rstudio.github.io/shinytest/articles/package.html) though a reason is not specified.

It reports that the coverage is only 27% I would aim for 90% or more seeing that you want to test also the shiny app and the code.

> 6. A lot of the code is dedicated to the interactive command line terminal, and checking for errors in the user responses. Since it is interactive, this code cannot be tested with testthat. As a result, I've added #nocov tags around the command prompt lines to exclude them from testing. There is no integration of shinytest (which I am using to test the shinyapp) and covr at the moment (see issue [here](https://github.com/r-lib/covr/issues/277)) which means the app has 0% coverage. So by excluding the command prompt lines and the shinyapp during coverage testing, I have ~97% coverage of the remaining code on codecov.

It also warns about using T and F instead of TRUE and FALSE.

> 7. Changed T and F in vignettes and also in other places (https://github.com/yihanwu/baRcodeR/commit/97fc17f5e71e15e8317f0c86c6523f8b205dcf61#diff-43d20606d79ec280d9646a3925c1cdd1)

Also it recommends to think about the dependency to qrcodes that is on Depends instead of Imports. I think it is sensible to have it Depends as without it this packages seems useless.

> 8. Yes, the reason that qrcode is a dependency is that the underlying data matrix for generating the qrcode in the package is unexported. I wanted to avoid copying the data matrix into this package so using Depends is necessary.

However, goodpractie::gp couldn't continue with the analysis due to some error with linter. Probably this is due to some long string of code.

> 9. There is a for loop in the function to generate the pdf which goodpractice::gp does not like. The reason this is needed is explained in a comment below . (Response 32)

**Check of the metadata**

The README doesn't have any badge of status, from repostatus.org. See this section

> 10. repostatus badge added (https://github.com/yihanwu/baRcodeR/commit/2979d3382043aa6a1cf913d9e6d737a7818919e9)

Most if not all of the results that the spelling check reports are from examples or names. Consider adding them the dictionary so that they are not reported again.

> 11. added WORDLIST in inst/ for spelling. (https://github.com/yihanwu/baRcodeR/commit/2979d3382043aa6a1cf913d9e6d737a7818919e9)

The readme recommends to provide the sessionInfo, you might be interested on the package reprex, which can help your users to create the minimal reproducible example.

> 12. I've converted the README.md to be generated from a README.Rmd file. (https://github.com/yihanwu/baRcodeR/commit/2979d3382043aa6a1cf913d9e6d737a7818919e9)

The DESCRIPTION lists 6 Imports and 1 Depends but the NAMESPACE only imports two package.
The NAMESPACE should import fully the Depends packages and selectively the functions of packages on Imports.

> 13. I'm slightly confused about the NAMESPACE vs DESCRIPTION imports. Right now, I'm listing packages in the DESCRIPTION imports and referring to them with packagename::fun() like the R Packages book suggests [here](http://r-pkgs.had.co.nz/namespace.html).

It also uses Authors@R and Maintainer it is better to select just one (less things to keep up to date). Nice addition of the ORCID numbers.

> 14. If the maintainer field is removed, the maintainer is automatically generated from Authors@R to reference the creator. Right now though, Robert Colautti is the creator, but the person maintaining the project is me (@yihanwu).

I would recommen to use usethis::use_tidy_description to set the fields of the DESCRIPTION in a "standard" format.

> 15. The DESCRIPTION file has been regenerated using usethis. (https://github.com/yihanwu/baRcodeR/commit/2979d3382043aa6a1cf913d9e6d737a7818919e9)

**Check of the documentation**

The first thing I notice looking at the documentation is a link to another library.
I would move this reference below "Usage from the console".

> 16. This has been done (https://github.com/yihanwu/baRcodeR/commit/d60c7d3d53bb7c44f13ceeabe73f92daa04b7725)

I would also have a short section showing how you can use the library on the README.Rmd via the command line (without prompts). This way the potential user get a grasp of what does the package.

> 17. A quick example has been added as the first section in the vignette after the overview. (https://github.com/yihanwu/baRcodeR/commit/d60c7d3d53bb7c44f13ceeabe73f92daa04b7725).

Also probably it would be nice to show the final output of the package, currently it seems like the result is a plain data.frame.

> 18. I've included an example page (using default layout) as part of the example from the above line. (https://github.com/yihanwu/baRcodeR/commit/d60c7d3d53bb7c44f13ceeabe73f92daa04b7725)

The section in the README about "Using the RStudio addin" I think it worth to be in its own vignette.

> 19. I've made a new vignette and shortened the README length (https://github.com/yihanwu/baRcodeR/commit/92be54710607222419c50b16bfc73bed4bc55f6a)

There are some man pages (man/cheatsheet.Rd, man/make_labels.Rd) that call library or require on baRcodeR, it is not necessary.

> 20. I've removed the call to library for cheatsheet but for make_labels, the call is necessary due to the export issue with qrcode explained above. (https://github.com/yihanwu/baRcodeR/commit/8e440710bcd1efb2b0d79d00852a1015f5ac7a60#diff-53e56b637346c7419dd7637bdd7c8c07)

It is not recommended to call the system on a man page like in create_PDF.Rd. If you have to, use system, but system2 would be better because the command is always quoted. So if you want to pass several commands you should call several times to system2.

> 21. system has been changed to system2. (https://github.com/yihanwu/baRcodeR/commit/8e440710bcd1efb2b0d79d00852a1015f5ac7a60#diff-53e56b637346c7419dd7637bdd7c8c07)

There is no information about what do the cheatsheet and the make_labels return.
It is always good to tell your users what to expect from your functions.

> 22. Added return values to documentation (https://github.com/yihanwu/baRcodeR/commit/8e440710bcd1efb2b0d79d00852a1015f5ac7a60#diff-53e56b637346c7419dd7637bdd7c8c07)

There isn't any example on custom_create_PDF, I would recommend to add one even if they are not run automatically.

> 23. Copied examples for create_PDF to custom_create_PDF. (https://github.com/yihanwu/baRcodeR/commit/8e440710bcd1efb2b0d79d00852a1015f5ac7a60#diff-53e56b637346c7419dd7637bdd7c8c07)

Also some examples do not run. From create_PDF:

example_vector <- as.data.frame(c("ao1", "a02", "a03"))
create_PDF(Labels = example_vector, name = file.path(tempdir(), "example"))
#   Cannot find a label column. Using first column as label input.
The details section of create_PDF has also a "#` " at the beginning.
The roxygen2 notation is duplicated on the source file, remove it and it will disappear

> 24. Fixed non-working example. Fixed extra roxygen. (https://github.com/yihanwu/baRcodeR/commit/8e440710bcd1efb2b0d79d00852a1015f5ac7a60#diff-53e56b637346c7419dd7637bdd7c8c07)

### Check the vignette

The image of the workflow is very informative and useful for the user, but I would move it below the overview (as well as the cheatsheet).

> 25. Done. (https://github.com/yihanwu/baRcodeR/commit/d60c7d3d53bb7c44f13ceeabe73f92daa04b7725)

As a user I would like to know what does the package first and then how can I do it.

The first code block of the vignette is to install some packages that are (or should be) automatically installed when the user installs baRcodeR.
The code block should be removed.

> 26. Done. (https://github.com/yihanwu/baRcodeR/commit/d60c7d3d53bb7c44f13ceeabe73f92daa04b7725)

In the vignette you use T and F instead of TRUE and FALSE, T and F are not protected, so it might introduce errors, change them to the full name please.

> 27. T's and F's have been replaced in the vignette. See commit on this from above . (Response 7)

Next you explain the workflow using prompts. Prompts are quoted, as they are printed in the terminal I would show them as code by using the three backticks formatting.

> 28. I've used backticks to format the prompts as a monospace font to reflect that they are on the command line. In addition, using them as quotes differentiates these lines from commands that should be entered directly on the command lines. (https://github.com/yihanwu/baRcodeR/commit/97fc17f5e71e15e8317f0c86c6523f8b205dcf61#diff-43d20606d79ec280d9646a3925c1cdd1)

The next section is again using prompts. I was a bit confused about what is the goal of this section.
The columns of this hierarchical labels are the same of the text, which might be confusing.
Maybe the title could be one and the string used a different.

> 29. In the vignette,  I've added a clarification that the result of following the "user prompt" section and the "by argument" section for each function should end in the same result. (https://github.com/yihanwu/baRcodeR/commit/97fc17f5e71e15e8317f0c86c6523f8b205dcf61#diff-43d20606d79ec280d9646a3925c1cdd1)

In the section "By argument" when I ran the code I got a warning:

#   Digits specified less than max level number. Increasing number of digits for level
It is mentioned again to store it in a csv. I tried to create_PDF and I didn't know where the file was created.

Apparently it is created on the working directory.

Maybe it should be better explained on the help page.

> 30. I've added a second warning to the return value on the help page and emphasized the file naming with an example as part of the parameter explanation. (https://github.com/yihanwu/baRcodeR/commit/97fc17f5e71e15e8317f0c86c6523f8b205dcf61#diff-43d20606d79ec280d9646a3925c1cdd1)

When I used the argument user = TRUE in create_PDF I got a prompt, I introduced the name of the pdf with quotes. Then the resulting file got also quotes. This could be checked and warned to the user as it could create problems.

> 31. I've added a warning on the vignette for users to avoid quotes when typing strings into the interactive prompts. (https://github.com/yihanwu/baRcodeR/commit/97fc17f5e71e15e8317f0c86c6523f8b205dcf61#diff-43d20606d79ec280d9646a3925c1cdd1)

### Check the functionality

The functions are very well commented about what do they do.

Almost all the functions implement a sort of prompt.

I am not sure this is actually helpful as then the user need to know which question corresponds with which argument of the function.

The creation of labels seems a bit over-complicated to me.

We can obtain all the combinations of a group of variables with expand.grid

labels <- expand.grid(fixed = "a", variable = 1:5, join = "-",
                      text = letters[2:3], numbers = 7:9)
labels
apply(labels, 1, paste0, collapse = "")
This allows to omit the loop if you have a hierarchy.

> 32. The loop is necessary in this case because the level number of the labelling hierarchy is variable in each instance and therefore we cannot define the dimensions of expand.grid beforehand.

The GUI addin is nice, but the "Simple ID Code Generation" and the "Hierarchical ID Code Generation" end up creating a labels.csv instead of creating the bar codes.

I expected that they would be directly created.

> 33. There were two reasons that we separate ID Code Generation from PDF Creation. First, creating label.csv is important for reproducibility, especially if the user forgot to save the code snippet that was created or the steps in generating the labels. Saving the csv reinforces a reproducible approach and also provides a file that can archived or exported to other pipelines. Second, if the user has their own custom labels, they can skip the ID Code Generation step and go straight to creating the PDF without having to worry about generating the labels again.  This is also important for reproducibility since re-generating the codes from scratch increases opportunities for error.

Also there isn't a way to select where the labels.csv is created.

> 34. The labels save to the working directory. See comments below  for the commit where I added text notifications to the addin. (Response 49)

I tried to create a Hierarchical ID Code Generation with the GUI and I end up with an error :

Error: Input list has only one level. Did you forget a level or are you sure you are not looking for uniqIDMaker()?
That's after creating a label and pressing "Add level".

> 35. This was a result of shiny evaluating the input in real-time and converting the warning in uniqID_hier_maker into an error. If there is only one level then the user should use the "Simple ID creation" tab. If another level is added, then the error disappears. However, I can see how this could cause confusion so the prompt "Please add a level" was added and the preview of labels will not show up until the user adds two levels. (https://github.com/yihanwu/baRcodeR/commit/3561ba66267d20fa16a50f9433c9f4a40f9cd8a2)

### Inspection of the code

Whenever possible avoid sapply(); use vapply(), as sapply will simplify the output which might cause errors. There are 8 cases of this, maybe some of them can be replaced, in some cases you just want to get the first element, maybe you can use unlist instead.

Also you use 1:... like 5 times; use seq_len() or seq_along() or simply seq().

> 36. Those have been replaced in the code with seq(1, ...) (https://github.com/yihanwu/baRcodeR/commit/f6678cf57fdd6963035b3c3b0f4b28795cb7ac19)

In code_128_make() you used T and F replace them for TRUE and FALSE.

> 37. Replaced T and F with TRUE and FALSE.

In hidden_createPDF.R you check a class with class(), it is better to use the is() function as it will show all the classes it inherits from.

> 38. With the class check, I only want the first "class" value to determine if the input is a vector or data.frame object and having the test fail if it is neither. It's a simple and quick check for incorrect input formats. Having all the classes of an object (as returned by the is.X function) would make it a less efficient test since I would have to check each class of an object against a list.

There are very long functions (make_labels_internals, custom_create_PDF and server that are more than 100 lines) try splitting them in shorter functions. Usually functions of less of 50 lines are recommended.

> 39. make_labels_internals is a long function because it contains the user interface. This includes elements for all three tabs as well as the shiny code needed to perform the required processes such as the preview image. custom_create_PDF is also long because more than half of the function is text for interactive user prompts and input checking. While I managed to split the actual generation of the barcodes into their own functions, the layout uses loops to generate a new page in the pdf file.

There are also very long lines, consider shorter lines; 253 lines (15%) are > 80 characters long.

> 40. Very long lines have been modified to become shorter lines for the three functions mentioned above (under 120) except for the warning and stop lines where new lines within the messages would break up the message printed for the user (https://github.com/yihanwu/baRcodeR/commit/f6678cf57fdd6963035b3c3b0f4b28795cb7ac19#diff-53e56b637346c7419dd7637bdd7c8c07).

The create_PDF and the custom_create_PDF do the same.

I would delete create_PDF.
I understand that the idea was to avoid confusing the user with too many arguments.
But either do not show them or let them trust you with the defaults, which work reasonable well.

> 41. We understand the logic here but since the package is already published on CRAN (and downloaded > 7,000 times), we think it would cause more confusion to change the function names at this point.

You make use of grDevices:: and other packages but don't declare on the NAMESPACE that you import them.
You should do so with #' @importFrom grDevices pdf for instance.

> 42. The use of functions not declared in the NAMESPACE has been addressed above . (Response 12)

Check the testing

The file uniqID_maker_addin.R is not tested.

> 43. uniqID_maker_addin.R is checked by shinytest, which does not integrate with covr and so it does not show up even though it is tested.

I don't know how can you test when you interactively ask the user with prompts but that section should be tested too (hidden_createPDF.R).

> 44. R does not have a framework for testing prompts and at this point, the best option seems to be manual testing of the prompts.

### Conclusions

Overall the package has its audience it is easy to use and provides a nice functionality for wet-dry researchers.

## raynamharris review

### Package Review

[…]
