### 2019-10-28 #338 @llrs
Great! Many thanks @yihanwu for answering all the points. The package has really improved  with all these commits.

1. & 2. Great!

3. I'm not sure that is the solution. Surely this makes that the warnings do not appear but it doesn't solve the issue.

5. Oh, it makes sense. Thanks for the links.

6. & 44. Here is an [example on how to test](https://stackoverflow.com/questions/41372146/) use input from the command line.

7 & other points. Great, sorry to be so repetitive.

8. I'm not sure I understood your response. But given that it works and that it is accepted by CRAN I won't check it further.

9 & 32. I can't say for sure, but it is strange that `goodpractice::gp` couldn't check a loop. I think this indicates some concerns with the package.

12. Great!

13. Ok, the second version of the book, in the [Imports section](https://r-pkgs.org/namespace.html#imports), recommends your method approach too.

14. The normal approach is to use auth and cre. So if Robert Colautti  is the author then auth and you use then cre.

18. Nice picture, now I get much better an idea of what I get.

19. In the new vignette there is some configuration code that it is shown.

28. Thanks

32. In general it is recommended to avoid loops. An example on what kind of setup would be impossible/hard to do without this loop would convince me.

33. Yes, I understand your point but I think that if a users needs to use the GUI it can be all done in a single step, both save the labels.csv and create the labels.

39. In general more concise functions help to maintain the code.

40. One of the few lines I saw can be further simplified using is.numeric:

```
any(
    unlist(
      lapply(c(
        numcol, numrow,
        Fsz, ERows, ECols,
        trunc, page_width, page_height,
        height_margin, width_margin,
        x_space, y_space), class)
      ) != "numeric") == TRUE)
```
for

```
all(vapply(c(
        numcol, numrow,
        Fsz, ERows, ECols,
        trunc, page_width, page_height,
        height_margin, width_margin,
        x_space, y_space), is.numeric, logical(1L))
```

Note that you don't need to first check for the class then check if the class is the one you need and then compare if any is true. For this short loops it might not matter but for other things it can help you to speed up the process.

I also think that some warnings can be split in two sentences. One the consequences and another for the advice. For instance, "Linear barcodes created will have bar width smaller than 0.03 inches which may be unreadable by some barcode scanners."
can be plitted into "Linear barcodes created will have bar width smaller than 0.03 inches.","\nIncrease the label_width to make them readable."
