### 2019-10-29 #338 @yihanwu
Hi @llrs,

I just wanted to get clarifications on some of the points before I start making the changes. These are numbered based on your comments.

4. In the example referenced here, the solution is to set a connection and read in a test response for a function with one user prompt. I think it is possible to get this to work with something like this ...

```
library(testthat)

options(mypkg.connection = stdin())

marryme <- function() {
  cat("will you marry me? (y/n) > ")
  ans <- readLines(con = getOption("mypkg.connection"), n = 1)
  cat("\n")
  cat("will this work? (y/n) >")
  ans1 <- readLines(con = getOption("mypkg.connection"), n = 1)
  cat("\n")
  cat("will this work x2? (y/n) >")
  ans2 <- readLines(con = getOption("mypkg.connection"), n = 1)
  cat("\n")
  cat("will this work x3? (y/n) >")
  ans3 <- readLines(con = getOption("mypkg.connection"), n = 1)
  cat("\n")
  return(ans3=="n")
}

test_that("input", {

  f <- file()
  options(mypkg.connection = f)
 # two sets of four answers
  ans <- paste(c("n", "y", "y", "n", "n", "y", "y", "y"), collapse = "\n")
  write(ans, f)

  expect_true(marryme())
  expect_true(marryme())

  options(mypkg.connection = stdin())

  close(f)
})
```
This is enormously helpful if I can get it to work with all the possible test cases.

18.  For the large loop, here it is in basic steps. I've struggled with trying to vectorize this so any suggestions will be greatly appreciated.

1. vector of raster image barcodes is generated with lapply (these come without the text string)
2. make a table of viewports (based on parameters). Each cell will end up with a viewport for the barcode, and the text string in the end.
3. The loop over the vector of barcodes:
a. find x position and y position of the cell (carried over from last loop after the first iteration where starting position is determined by row and column skipping paramters) in the table to put barcode in
b. if x position and y position  is beyond limits of the table rows and columns,  start a new page with a new grid
c. go into the cell and put in the raster barcode and string in two viewports, moving across and down the table as required.
iv. x position and y position will be incremented and these values determine the values used in step a.

I'm not sure how I can open a new page when needed without having the information from the previous loop regarding where the next barcode needs to go.

19. How about both "Create Labels" and "Create Labels + PDF" buttons so the user can have a choice of either and go to the PDF creation tab if they need customized PDFs.
