### 2019-10-30 #338 @llrs
Hi @yihanwu

4. Glad it helped. Sorry I can't help you more on this, but I don't have experience on testing input from command line.

18. The positions on a page are fixed (unless the grid changes in each page, which doesn't seem the case). So, once you know them for a page you can calculate how many pages you need and assign the barcodes to each position and page. If you have each x, y and page position on a data.frame in a single loop you can calculate the image barcode from the labels and use the position to generate the pdf.

This is not needed to be accepted but I think it can be helpful if the users needs to print several hundreds of barcodes (or for future tasks you start).

19. Yes, that will be helpful. Now that I saw your idea, I realized that my confusion was with the text on the buttons. To me the labels are on the pdf to be printed and used while the buttons considered labels as the matrix with the text.
