### 2019-11-07 #338 @yihanwu
More updates:

1. I've created tests for user prompts by using several internal functions. (https://github.com/yihanwu/baRcodeR/commit/3abede8d5cfba70c7cf58d24a77a3408ff5257e5). The internal functions allow for several tries at putting in the correct response, and also a choice selection similar to the `menu()` function (unfortunately `menu()` did not seem to work for testing)

2. The loop problem with goodpractice disappeared after I simplified the user prompts.

3. I've added to the note in the add-in directing the user to head to the barcode creation tab. Unfortunately, when I tried out the extra button to "create labels and PDF", it appeared even more confusing. (https://github.com/yihanwu/baRcodeR/commit/ea73d05d733c227c2d3629f65bccc4f70ad5a6a0)

3. PDF generation now generates the locations of each label first, and the loop is simpified for plotting with grid. (https://github.com/yihanwu/baRcodeR/commit/fdabc9f95d54abbc9275ac2c8ad1d12347825048)

4. The description has been changed so the maintainer field is automatically generated. (https://github.com/yihanwu/baRcodeR/commit/fdabc9f95d54abbc9275ac2c8ad1d12347825048)
