### 2019-11-26 #338 @yihanwu
Hi @sckott, @llrs , @raynamharris,

Are there any other changes I should review and make?

Cheers,
Yihan
