### 2019-11-27 #338 @yihanwu
Hi @sckott,

I've made the changes you have suggested but when I go to transfer the repository, Github tells me I "do not have permission to create public repositories". Is there something else I need to do first?
