### 2019-11-27 #338 @sckott
you should have got an email invitation to a github team within ropensci, did you not get that?
