### 2019-11-27 #338 @yihanwu
Ah, I found it. Completely my fault.

The repo transferred successfully. I've added the docs url pointing to docs.ropensci and removed the automatic deployment from Travis to gh-pages.

I tried to access the generated docs at https://docs.ropensci.org/baRcodeR. Is there a delay until they should be up?
