### 2019-11-28 #338 @sckott
its up now https://docs.ropensci.org/baRcodeR/
